+++ source/limit_library_search_by_file_extension/changelog.md
@@ -1,4 +1,8 @@
 
+**<span style="color:#56adda">0.0.3</span>**
+- Add ability to force all matching files to be added to the pending task queue
+- Remove support for Unmanic v1 PluginHandler compatibility
+
 **<span style="color:#56adda">0.0.2</span>**
 - Update Plugin for Unmanic v2 PluginHandler compatibility